01_authentication/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
rsa
jwt